Extract "File Tree" control from FormBrowse
